Changed recipe format, +quantity rounding
Recipe item quantity is now a string, allowing for fractional values.
This is in anticipation of adding a field specifying the measurement
type, which will include plenty of e.g. 3/4 cup.

The Planner will now also round the quantity values up to the nearest
integer when calculating the shopping list.  User interface only allows
modifications by ints as well.