+some view stuff
View functionality is coming along. The infrastructure is in place  for
the remainder. Printing/updating the screen is implemented, along with
the framework for menu navigation.

As-is a shallow proof of concept allows user to enter select recipe
list, select/deselect the recipes, with accompanying color changing
(persistent, too!), move back to the main menu, then return to select
recipes and see the persistent color.