+Shopping list, cart population
Users can now create a shopping list from the selected recipes. This
shopping list is treated as a recipe, so it is able to edited in the
same way a normal recipe is, allowing users to customize the final list
as desired.

Main Menu's "Place Order" Has been replace with "Fill Cart", allowing
the user final review in a web browser.  Scraping with selenium is
sluggish and perhaps not the approach to use to relay curbside pickup
timeslots to user.  This could still work just fine for grocery
delivery.

The big brain stuff is incorporating the idea of the grocery list being
just a recipe, though this requires the planner to translate a
dictionary structure.  The view's presentation of the grocery "recipe"
needs to be improved - it currently prints the str representation of the
"recipe_item" dictionary.