+token refresh, saving token
Communicator now checks that the access token is valid before making API
calls (though it doesn't recover upon "invalid token" responses).  If
the token is expired (as far as its reckoning goes) then it will use the
refresh token to acquire new tokens.

Refresh tokens are now saved to disk upon reciept.  This is read on
bootup to minimize the need for "human" interaction with selenium.  Much
nicer. And fewer IP bans!