Further understanding of mousewheel scrolling
Encountered unexpected behavior binding <Enter> and <Leave> to frames.
Seems that the highest/top widget is considered with this, so a frame 1
fit perfectly over frame 2 will be what's counted, not frame 2. If frame
1 covers half of frame 2, the scrolling will work over the half of frame
  2 that isn't covered (assuming <Enter> and <Leave> are bound to frame
    2).

So, the "deepest" widgets will need to have a reference to the canvas in
the scroll frame.