Further scrollwheel progress
The top widget gets the "focus" and must have the appropriate <Enter>
and <Leave> bindings in place to have effect. Current iteration has all
the highest-placed widgets (left side pane) with the binding and so
scrolls no matter where the cursor is (as long as it's in the
canvas_window).